streamlit_ui/requirements.txt
streamlit
google-cloud-storage
requests
